tests/wpt/css-tests/css-ui-3_dev/html/caret-color-021.htm

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>CSS Basic User Interface Test: Default caret-color test animation</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas">
<link href="http://www.w3.org/TR/css3-ui/#caret-color" rel="help">
<link href="https://www.w3.org/TR/css3-animations/#animation-shorthand-property" rel="help">
<link href="https://www.w3.org/TR/css3-color/#color0" rel="help">
<meta content="Test checks that the default value for caret-color property, which is 'auto', is not interpolable." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @keyframes caret-color-to-lime {
    to { caret-color: lime; }
  }

  #textarea {
    color: magenta;
    animation: caret-color-to-lime 2s -1s paused;
  }
</style>
</head><body>
  <textarea id="textarea"></textarea>
  <div id="log"></div>

  <script>
    test(
      function(){
        var textarea = document.getElementById("textarea");
        assert_equals(getComputedStyle(textarea).caretColor, 'rgb(0, 255, 0)');
      }, "Default caret-color is not interpolable");
</script>


</body></html>
